make extract metadata from string common
add title metadata support for vitest
fix problem with stucked fixtures
